Remark refactor
