+++ packages/mdx/src/remark/editor.tsx
@@ -1,10 +1,11 @@
 import { visitAsync } from "./unist-utils"
 import { transformEditor } from "./code"
 import { JsxNode, SuperNode } from "./nodes"
+import { CodeHikeConfig } from "./config"
 
 export async function transformEditorNodes(
   tree: SuperNode,
-  { theme }: { theme: any }
+  config: CodeHikeConfig
 ) {
   await visitAsync(
     tree,
@@ -13,7 +14,7 @@ export async function transformEditorNodes(
       if (node.name === "CH.Code") {
         await transformEditor(
           { node, index, parent: parent! },
-          { theme }
+          config
         )
       }
     }
