+++ packages/mdx/src/remark/scrollycoding.ts
@@ -4,13 +4,14 @@ import {
   CH_CODE_CONFIG_PLACEHOLDER,
 } from "./unist-utils"
 import { extractStepsInfo } from "./steps"
-import { getPresetConfig } from "./preview"
+import { getPresetConfig } from "./transform.preview"
 import { transformLinks } from "./section"
 import { SuperNode } from "./nodes"
+import { CodeHikeConfig } from "./config"
 
 export async function transformScrollycodings(
   tree: SuperNode,
-  config: { theme: any }
+  config: CodeHikeConfig
 ) {
   await visitAsync(
     tree,
@@ -24,11 +25,11 @@ export async function transformScrollycodings(
 }
 async function transformScrollycoding(
   node: SuperNode,
-  { theme }: { theme: any }
+  config: CodeHikeConfig
 ) {
   const editorSteps = await extractStepsInfo(
     node,
-    { theme },
+    config,
     "merge step with previous"
   )
