+++ packages/mdx/src/remark/steps.tsx
@@ -1,13 +1,14 @@
 import { EditorStep } from "../mini-editor"
 import { isEditorNode, mapAnyCodeNode } from "./code"
 import { reduceSteps } from "./code-files-reducer"
+import { CodeHikeConfig } from "./config"
 import { SuperNode } from "./nodes"
 
 // extract step info
 
 export async function extractStepsInfo(
   parent: SuperNode,
-  config: { theme: any },
+  config: CodeHikeConfig,
   merge:
     | "merge steps with header"
     | "merge step with previous"
